+++ vite.config.ts
@@ -1,32 +1,25 @@
 import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-import dts from 'vite-plugin-dts';
 import { resolve } from 'path';
+import react from '@vitejs/plugin-react';
 
 export default defineConfig({
-  plugins: [
-    react(),
-    dts({
-      include: ['src'],
-      exclude: ['src/**/*.test.ts', 'src/**/*.spec.ts'],
-    }),
-  ],
   build: {
     lib: {
       entry: resolve(__dirname, 'src/index.ts'),
-      name: 'ExpoAesVetkeys',
-      fileName: 'index',
       formats: ['es'],
     },
     rollupOptions: {
-      external: ['react', 'react-native', '@dfinity/agent'],
-      output: {
-        globals: {
-          react: 'React',
-          'react-native': 'ReactNative',
-          '@dfinity/agent': 'Agent',
-        },
-      },
+      external: [
+        'react',
+        'react-native',
+        '@dfinity/agent',
+        '@react-native-async-storage/async-storage',
+        'expo-secure-store',
+        'expo-crypto',
+        'expo-crypto-universal',
+        'expo-storage-universal',
+      ],
     },
   },
+  plugins: [react()],
 });